refactor(di): isolate lifecycle reprojection wiring
