+++ src/backend/src/agentclaw/community/di/modules/skill_center_group4_module.py
@@ -33,6 +33,28 @@ from agentclaw.community.core.repository.protocols.skill_center_reference import
 from agentclaw.community.core.repository.protocols.track_latest import (
     TrackLatestRepositoryProtocol,
 )
+from agentclaw.community.core.repository.protocols.bot import BotRepository
+from agentclaw.community.core.repository.protocols.devices import DeviceBindingRepository
+from agentclaw.community.core.repository.protocols.skills_pool import (
+    SkillsPoolLayoutRepositoryProtocol,
+)
+from agentclaw.community.core.devices.services.device_context_resolver import (
+    DeviceContextResolver,
+)
+from agentclaw.community.core.skill_center.services.lifecycle_runtime_reprojection import (
+    LifecycleRuntimeProjectionTaskHandler,
+    LifecycleRuntimeProjectionWakeup,
+)
+from agentclaw.community.core.skill_center.services.mcp_runtime_probe import (
+    CurrentMcpRuntimeProbeService,
+)
+from agentclaw.community.core.skill_center.services.runtime_projections.registry import (
+    EngineRuntimeProjectionRegistry,
+)
+from agentclaw.community.core.skills_pool.types import (
+    BotSkillLayoutScope,
+    runtime_uses_pool_paths,
+)
 from agentclaw.community.core.skill_center.services.group4_task_registrar import (
     SkillCenterGroup4TaskRegistrar,
 )
@@ -57,6 +79,9 @@ from agentclaw.community.core.skill_center.services.track_latest_event_listener 
 from agentclaw.community.core.task_queue.services.registry import HandlerRegistry
 from agentclaw.community.core.task_queue.services.task_queue_service import TaskQueueService
 from agentclaw.community.plugin_api.cache import CachePlugin
+from agentclaw.community.plugin_api.device_adapter_transport import (
+    DeviceAdapterTransport,
+)
 from agentclaw.community.core.skill_center.skill_center_gateway_service_protocol import (
     SkillCenterGatewayServiceProtocol,
 )
@@ -73,6 +98,73 @@ class SkillCenterGroup4Module(Module):
             TrackLatestRepositoryProtocol,
             to=TrackLatestRepository,
             scope=singleton,
+        )
+
+    @singleton
+    @provider
+    def current_mcp_runtime_probe_service(
+        self,
+        resolver: DeviceContextResolver,
+        adapter_transport: DeviceAdapterTransport,
+    ) -> CurrentMcpRuntimeProbeService:
+        return CurrentMcpRuntimeProbeService(
+            resolver=resolver,
+            adapter_transport=adapter_transport,
+        )
+
+    @singleton
+    @provider
+    def lifecycle_runtime_projection_task_handler(
+        self,
+        binding_repository: DeviceBindingRepository,
+        bot_repository: BotRepository,
+        projection_registry: EngineRuntimeProjectionRegistry,
+        runtime_reconciler: BotRuntimeProjectorProtocol,
+        mcp_probe: CurrentMcpRuntimeProbeService,
+        layout_repository: SkillsPoolLayoutRepositoryProtocol,
+    ) -> LifecycleRuntimeProjectionTaskHandler:
+        def skill_projection_owned_elsewhere(bot: dict[str, object]) -> bool:
+            if bot.get("bot_type") != "desktop":
+                return False
+            values = (bot.get("env"), bot.get("entity_id"), bot.get("bot_id"))
+            if not all(isinstance(value, str) and value for value in values):
+                return False
+            state = layout_repository.get(
+                BotSkillLayoutScope(
+                    env=str(values[0]),
+                    entity_id=str(values[1]),
+                    bot_id=str(values[2]),
+                )
+            )
+            return runtime_uses_pool_paths(state)
+
+        return LifecycleRuntimeProjectionTaskHandler(
+            binding_repository=binding_repository,
+            bot_repository=bot_repository,
+            projection_registry=projection_registry,
+            projector=runtime_reconciler,
+            mcp_probe=mcp_probe,
+            skill_projection_owned_elsewhere=skill_projection_owned_elsewhere,
+        )
+
+    @singleton
+    @provider
+    def lifecycle_runtime_projection_wakeup(
+        self,
+        binding_repository: DeviceBindingRepository,
+        bot_repository: BotRepository,
+        task_queue_service: TaskQueueService,
+        projection_registry: EngineRuntimeProjectionRegistry,
+        registry: HandlerRegistry,
+        task_handler: LifecycleRuntimeProjectionTaskHandler,
+    ) -> LifecycleRuntimeProjectionWakeup:
+        return LifecycleRuntimeProjectionWakeup(
+            binding_repository=binding_repository,
+            bot_repository=bot_repository,
+            task_queue_service=task_queue_service,
+            projection_registry=projection_registry,
+            registry=registry,
+            task_handler=task_handler,
         )
 
     @singleton
